+++ src/main/java/name/divinityunbound/block/ModBlocks.java
@@ -95,6 +95,9 @@ public class ModBlocks {
     public static final Block SPACE_TIME_AMALGAMATOR = registerBlock("space_time_amalgamator",
             new SpaceTimeAmalgamatorBlock(FabricBlockSettings.copyOf(Blocks.IRON_BLOCK).nonOpaque()));
 
+    public static final Block WORMHOLE_TRANSPORTER = registerBlock("wormhole_transporter",
+            new WormholeTransporterBlock(FabricBlockSettings.copyOf(Blocks.IRON_BLOCK).nonOpaque()));
+
 
     private static Block registerBlock(String name, Block block) {
         registerBlockItem(name, block);