+++ src/main/java/name/divinityunbound/datagen/ModBlockTagProvider.java
@@ -40,7 +40,8 @@ public class ModBlockTagProvider extends FabricTagProvider.BlockTagProvider {
                 .add(ModBlocks.QUANTITY_UPGRADE)
                 .add(ModBlocks.UNHOLY_SILENCER)
                 .add(ModBlocks.SPACE_TIME_EVAPORATOR)
-                .add(ModBlocks.SPACE_TIME_AMALGAMATOR);
+                .add(ModBlocks.SPACE_TIME_AMALGAMATOR)
+                .add(ModBlocks.WORMHOLE_TRANSPORTER);
 
         getOrCreateTagBuilder(BlockTags.NEEDS_IRON_TOOL)
                 .add(ModBlocks.CELESTITE_BLOCK)
@@ -57,7 +58,8 @@ public class ModBlockTagProvider extends FabricTagProvider.BlockTagProvider {
                 .add(ModBlocks.SPACE_SIPHON)
                 .add(ModBlocks.SPEED_UPGRADE)
                 .add(ModBlocks.QUANTITY_UPGRADE)
-                .add(ModBlocks.UNHOLY_SILENCER);
+                .add(ModBlocks.UNHOLY_SILENCER)
+                .add(ModBlocks.WORMHOLE_TRANSPORTER);
 
         getOrCreateTagBuilder(BlockTags.NEEDS_DIAMOND_TOOL)
                 .add(ModBlocks.SPACE_TIME_EVAPORATOR)